==> notebooks/converting/1.0-ajmk-installed-capacity.ipynb <==
import pandas as pd

installed_capacity = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/raw/starter-kits/kenya/Table1_KEN.csv")
installed_capacity

installed_capacity = installed_capacity.rename(columns={"Power Generation Technology": "Technology"})
installed_capacity.head()

latest_installed_capacity = installed_capacity[installed_capacity.Year == installed_capacity['Year'].max()]
latest_installed_capacity

technoeconomics = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/raw/starter-kits/kenya/Table2_KEN.csv")
technoeconomics

installed_capacity_cf = latest_installed_capacity.merge(technoeconomics[technoeconomics.Parameter == "Average Capacity Factor"], on="Technology")
installed_capacity_cf = installed_capacity_cf.rename(columns={"Value_y": "average_capacity_factor", "Value_x": "estimated_installed_capacity_MW"})
installed_capacity_cf = installed_capacity_cf.drop(columns=["Parameter_y", "Parameter_x"])
installed_capacity_cf

installed_capacity_cf['estimated_installed_capacity_PJ_y'] = installed_capacity_cf.estimated_installed_capacity_MW * 365 * installed_capacity_cf.average_capacity_factor
installed_capacity_cf

installed_capacity_pj_y = installed_capacity_cf.drop(columns=["estimated_installed_capacity_MW","average_capacity_factor"])
installed_capacity_pj_y

installed_capacity_pj_y_wide = installed_capacity_pj_y.pivot(index='Technology', columns='Year', values='estimated_installed_capacity_PJ_y').reset_index()
installed_capacity_pj_y_wide

unknown_cols = list(range(2020, 2055, 5))
for col in unknown_cols:
    installed_capacity_pj_y_wide[col] = 0
installed_capacity_pj_y_wide

installed_capacity_pj_y_wide.insert(1, "RegionName", "R1" )
installed_capacity_pj_y_wide.insert(2, "Unit", "PJ/y" )
installed_capacity_pj_y_wide

muse_installed_capacity = installed_capacity_pj_y_wide.rename(columns={"Technology":"ProcessName"}).set_index("ProcessName")
muse_installed_capacity

## Oil installed capacity

This is zero for all as there is no data available

oil_techno = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/raw/starter-kits/Kenya/Table5_KEN.csv")
oil_techno

oil_techs = list(pd.unique(oil_techno.Technology))
oil_techs

existing_capacity_dict = {}
for tech in oil_techs:
    existing_capacity_dict[tech] = [tech, "Kenya", "PJ/y"] + [0]*17
    
existing_capacity_dict

oil_existing_capacity = pd.DataFrame.from_dict(existing_capacity_dict, orient="index", columns = ['ProcessName', "RegionName", "Unit"] + list(range(2018, 2052, 2)))
oil_existing_capacity.reset_index(drop=True)

# list(range(2018, 2052, 2))
