==> notebooks/converting/3.0-ajmk-crude-oil.ipynb <==
import pandas as pd
pd.set_option('display.max_columns', None)

oil = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/raw/starter-kits/kenya/Table5_KEN.csv")
oil

oil["RegionName"] = "kenya"
oil["Time"] = "2020"
oil["Level"] = "fixed"
oil["cap_exp"] = 1
oil["fix_exp"] = 1
oil["var_par"] = 0
oil["var_exp"] = 1
oil["Type"] = "energy"
oil["EndUse"] = "oil"
oil["Agent2"] = 1
oil["InterestRate"] = 0.1
oil["MaxCapacityAddition"] = 20
oil["MaxCapacityGrowth"] = 20
oil["TotalCapacityLimit"] = 20

oil = oil.reset_index()
oil

oil_renamed = oil.rename(columns={"Capital Cost ($/kW in 2020)": "cap_par", "Operational Life (years)": "TechnicalLife", "Technology": "ProcessName"})
oil_renamed

oil_renamed = oil_renamed.drop(columns="Output Ratio")
oil_renamed

muse_technodata = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/external/muse_data/default/technodata/gas/Technodata.csv")
muse_technodata

oil_renamed["Fuel"] = "oil"
oil_renamed["efficiency"] = 1
oil_renamed["ScalingSize"] = 1
oil_renamed["UtilizationFactor"] = 1
oil_renamed["fix_par"] = 1

oil_renamed = oil_renamed.reindex(muse_technodata.columns, axis=1)
oil_renamed


oil_renamed = muse_technodata[muse_technodata.ProcessName == "Unit"].append(oil_renamed)
oil_renamed

oil_renamed.to_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/interim/technodata/oil_technodata.csv")

data_test = {}
data_test['power'] = {"technodata": 1}
data_test['power']['comm_in'] = 2

data_test['oil'] = {"technodata": 1}
data_test['oil']['comm_in'] = 2

data_test

for key in data_test:
    print(key)
