==> notebooks/converting/4.0-ajmk-CommIn-CommOut.ipynb <==
import pandas as pd

emissions = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/raw/starter-kits/kenya/Table7_KEN.csv")
emissions

emissions.Value *= 0.000001
emissions

CommIn = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/external/muse_data/default/technodata/gas/CommIn.csv")
CommIn

## Generate Oil CommOut

oil_technodata = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/interim/technodata/oil_technodata.csv")
oil_technodata

crude_oil_output = {
    "Crude Oil Refinery Option 1": {"oil": 1, "gas": 0, "electricity": 0, "CO2f": 0, "wind": 0},
    "Crude Oil Refinery Option 2": {"oil": 1, "gas": 0, "electricity": 0, "CO2f": 0, "wind": 0}
}
crude_oil_output

processes = list(pd.unique(oil_technodata[oil_technodata['ProcessName'] != "Unit"]['ProcessName']))
processes

units = {
    "ProcessName": ["Unit"],
    "RegionName": ["-"],
    "Time": ["Year"],
    "Level": ["-"],
    "electricity": ["PJ/PJ"],
    "gas": ["PJ/PJ"],
    "oil": ["PJ/PJ"],
    "wind": ["PJ/PJ"],
    "CO2f": ["kt/PJ"],
}

## Generate Power CommOut

power_technodata = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/interim/technodata/power_technodata.csv")
power_technodata

list(pd.unique(power_technodata[power_technodata.ProcessName != "Unit"]['ProcessName']))

power_fuels = list(pd.unique(power_technodata[power_technodata.ProcessName != "Unit"]['Fuel']))
power_fuels

example_technoeconomic = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/external/example_model/Techno_Economic.csv")
example_technoeconomic = example_technoeconomic.iloc[1:]
example_technoeconomic = example_technoeconomic.apply(pd.to_numeric, errors='ignore')
africa_technoeconomic = example_technoeconomic[(example_technoeconomic.RegionName=="OAFR") & (example_technoeconomic.Time<=2050)]
africa_technoeconomic['CommIn'] = 1/(africa_technoeconomic.GrossEfficiency/100)
africa_technoeconomic


technology_converter = {
    "Biomass Power Plant": "BIO_BIO",
    "CSP with Storage": "REW_CSP_NST",
    "CSP without Storage": "REW_CSP_NST",
    "Coal Power Plant": "COA_SUB_HC",
    "Gas Power Plant (CCGT)": "GAS_GT",
    "Gas Power Plant (SCGT)": "GAS_GT",
    "Geothermal Power Plant": "REW_GEO",
    "Large Hydropower Plant (Dam) (>100MW)": "REW_HYD_LG",
    "Light Fuel Oil Power Plant": "OIL_ST",
    "Light Fuel Oil Standalone Generator (1kW)": "OIL_ST",
    "Medium Hydropower Plant (10-100MW)": "REW_HYD_SM",
    "Nuclear Power Plant": "NUK_LWR",
    "Offshore Wind": "REW_WND_OFF",
    "Oil Fired Gas Turbine (SCGT)": "OIL_ST",
    "Onshore Wind": "REW_WND_ON",
    "Small Hydropower Plant (<10MW)": "REW_HYD_SM",
    "Solar PV (Distributed with Storage)": "REW_PV_GND",
    "Solar PV (Utility)": "REW_PV_GND"
}

power_types = power_technodata[power_technodata.ProcessName != "Unit"][["ProcessName", "Fuel"]].drop_duplicates()
power_types

power_types = power_technodata[power_technodata.ProcessName != "Unit"][["ProcessName", "Fuel"]].drop_duplicates()

power_types['ExampleTechs'] = power_technodata['ProcessName'].map(technology_converter)

power_types = power_types.merge(africa_technoeconomic[['ProcessName',  "CommIn", "Time"]], left_on="ExampleTechs", right_on="ProcessName")
power_types = power_types.drop(columns=["ProcessName_y", "ExampleTechs"])
power_types = power_types.rename(columns={"ProcessName_x": "ProcessName", "CommIn": "value"})
power_types = power_types.drop_duplicates()
power_types



power_types[power_types.ProcessName == "Coal Power Plant"]

power_comm_in = power_types.pivot(index=['ProcessName',"Time"], columns='Fuel', values="value").fillna(0)

power_comm_in.insert(0, "RegionName", "kenya")
power_comm_in.insert(1, "Time", 2020)
power_comm_in.insert(2, "Level", "fixed")
power_comm_in.insert(3, "electricity", 0)
power_comm_in['CO2f'] = 0

power_comm_in

pd.unique(power_types.Fuel)

emissions.Fuel = emissions.Fuel.str.lower()
emissions.Fuel = emissions.Fuel.str.replace("natural gas", "gas")
emissions.Fuel = emissions.Fuel.str.replace("crude oil", "oil")
emissions

power_types = power_technodata[power_technodata.ProcessName != "Unit"][["ProcessName", "Fuel"]].drop_duplicates()
power_types

power_types_emissions = power_types.merge(emissions.drop(columns="Parameter"), on="Fuel", how="left").fillna(0)
power_types_emissions = power_types_emissions.rename(columns = {"Value": "CO2f"})
power_types_emissions['value'] = 0
power_types_emissions

power_comm_out = power_types_emissions.pivot(index=['ProcessName', "CO2f"], columns='Fuel', values="value").fillna(0).reset_index()
power_comm_out['electricity'] = 1

power_comm_out.insert(1, "RegionName", "kenya")
power_comm_out.insert(2, "Time", 2020)
power_comm_out.insert(3, "Level", "fixed")

power_comm_out



power_comm_out_merged = pd.merge(power_comm_out, pd.Series(list(pd.unique(power_technodata.Time))[1:], name="Time"), how="cross")
power_comm_out_merged = power_comm_out_merged.drop(columns="Time_x")
power_comm_out = power_comm_out_merged.rename(columns={"Time_y":"Time"})
power_comm_out

units = {
    "ProcessName": ["Unit"],
    "RegionName": ["-"],
    "Time": ["Year"],
    "Level": ["-"],
    "CO2f": ["kt/PJ"],
    "biomass": ["PJ/PJ"],
    "coal": ["PJ/PJ"],
    "gas": ["PJ/PJ"],
    "geothermal": ["PJ/PJ"],
    "hydro": ["PJ/PJ"],
    "oil": ["PJ/PJ"],
    "solar": ["PJ/PJ"],
    "uranium": ["PJ/PJ"],
    "wind": ["PJ/PJ"],
    "electricity": ["PJ/PJ"],
}

units

units_row = pd.DataFrame.from_dict(units, orient="columns")
units_row
power_comm_out = units_row.append(power_comm_out)
power_comm_out

## Generate CommOut

def generate_comm_out(units, processes, outputs):
    if not all(process in outputs for process in processes):
        raise ValueError("Missing values in outputs dictionary: {}".format(outputs))
    else:
        comm_out = pd.DataFrame.from_dict(outputs, orient='index')

        comm_out.insert(0, "RegionName", "kenya")
        comm_out.insert(1, "Time", 2020)
        comm_out.insert(2, "Level", "fixed")
        
        comm_out = comm_out.reset_index()
        comm_out = comm_out.rename(columns={"index":"ProcessName"})

        units_row = pd.DataFrame.from_dict(units, orient="columns")
        comm_out = units_row.append(comm_out)

        return comm_out

generate_comm_out(units, processes, crude_oil_output)
