==> notebooks/converting/5.0-ajmk-agents-file.ipynb <==
import pandas as pd

agents = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/external/muse_data/default/technodata/Agents.csv")
agents

muse_data = {}

muse_data['inputs'] = {'GlobalCommodities': "global_commodities"}

muse_data["technodata"] = {"Agents": "agents"}

muse_data["technodata"]["power"] = {
    "ExistingCapacity": "existing_capacity"
}
muse_data["technodata"]["power"]["Technodata"] = "technodata"
muse_data["technodata"]["power"]["CommIn"] = "commin"
muse_data["technodata"]["power"]["CommOut"] = "commout"
muse_data["technodata"]["oil"] = {"Technodata": "technodata"}
muse_data["technodata"]["oil"]["CommIn"] = "commin"
muse_data["technodata"]["oil"]["CommOut"] = "commout"
muse_data["technodata"]["oil"][
    "ExistingCapacity"
] = "existingcaapcity"

muse_data

results_data = muse_data
for folder in results_data:
    for sector in results_data[folder]:
        if type(results_data[folder][sector]) is str:
            print(results_data[folder][sector])
        else:
            for csv in results_data[folder][sector]:
#             print(csv)
                print(csv)
