==> notebooks/converting/6.0-ajmk-GlobalCommodities.ipynb <==
import pandas as pd

commodities = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/raw/starter-kits/Kenya/Table7_KEN.csv")
commodities

commodities['Commodity'] = commodities['Fuel']
commodities = commodities.drop(columns="Parameter")
commodities['Fuel'] = commodities['Fuel'].str.lower().str.replace("light fuel oil", "LFO").str.replace("heavy fuel oil", "HFO").str.replace("crude oil","crude_oil").str.replace("natural gas", "gas")
commodities

commodities = commodities.rename(columns={"Value": "CommodityEmissionFactor_CO2", "Fuel": "CommodityName"})
commodities

commodities['CommodityType'] = "Energy"
commodities

commodities['HeatRate'] = 1
commodities['Unit'] = "kg CO2/GJ"
commodities

CO2_row = {"Commodity": "CO2fuelcomsbustion", "CommodityType": "Environmental", "CommodityName": "CO2f", "CommodityEmissionFactor_CO2":0, "HeatRate": 1, "Unit": "kt"}
commodities = commodities.append(CO2_row, ignore_index=True)
commodities

muse_commodities = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/external/muse_data/default/input/GlobalCommodities.csv")
muse_commodities

commodities = commodities.reindex(muse_commodities.columns, axis=1)
commodities

units = {
    "ProcessName": ["Unit"],
    "RegionName": ["-"],
    "Time": ["Year"],
    "Level": ["-"],
    "CO2f": ["kt/PJ"],
    "biomass": ["PJ/PJ"],
    "coal": ["PJ/PJ"],
    "gas": ["PJ/PJ"],
    "geothermal": ["PJ/PJ"],
    "hydro": ["PJ/PJ"],
    "HFO": ["PJ/PJ"],
    "LFO": ["PJx/PJ"],
    "crude_oil": ["PJ/PJ"],
    "solar": ["PJ/PJ"],
    "uranium": ["PJ/PJ"],
    "wind": ["PJ/PJ"],
    "electricity": ["PJ/PJ"],
}

for item in ["ProcessName", "RegionName", "Time", "Level", "CO2f", "crude_oil", "biomass", "coal", "LFO", "HFO", "gas"]:
    units.pop(item)

units["heat"] = "[PJ/PJ]"

for commodity, _ in units.items():
    entries = [commodity, "Energy", commodity, 0, 1, "kg C02/GJ"]
    new_row = {column: entry for column, entry in zip(list(commodities.columns), entries)}
    commodities = commodities.append(new_row, ignore_index=True)
    
commodities
