==> notebooks/converting/7.0-ajmk-market-projection.ipynb <==
import pandas as pd

costs = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/raw/starter-kits/Kenya/Table6_KEN.csv")
costs

import_costs = costs[~costs['Commodity'].str.contains("Extraction")]
import_costs

import_costs['Commodity'] = import_costs['Commodity'].str.replace("Imports", "")
import_costs['Commodity'] = import_costs['Commodity'].str.replace("Natural", "")
import_costs['Commodity'] = import_costs['Commodity'].str.lower()
import_costs['Commodity'] = import_costs['Commodity'].str.replace("light fuel oil", "LFO")
import_costs['Commodity'] = import_costs['Commodity'].str.replace("heavy fuel oil", "HFO")
import_costs['Commodity'] = import_costs['Commodity'].str.replace(" ", "")
import_costs['Commodity'] = import_costs['Commodity'].str.replace("crudeoil", "crude_oil")
import_costs

fuels = list(pd.unique(import_costs.Commodity))
fuels

projections = import_costs.pivot_table(index="Year", columns="Commodity", values="Value")
projections

projections['RegionName'] = "Kenya"
projections['Attribute'] = "CommodityPrice"
projections

projections = projections.reset_index()
projections = projections.rename(columns={"Year": "Time"})
projections

muse_projections = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/external/muse_data/default/input/Projections.csv")
muse_projections.head()


projections

col_order = ["RegionName", "Attribute", "Time"] + fuels
projections = projections[col_order]
projections

units = {
    "ProcessName": ["Unit"],
    "RegionName": ["-"],
    "Time": ["Year"],
    "Level": ["-"],
    "CO2f": ["kt/PJ"],
    "biomass": ["PJ/PJ"],
    "coal": ["PJ/PJ"],
    "gas": ["PJ/PJ"],
    "geothermal": ["PJ/PJ"],
    "hydro": ["PJ/PJ"],
    "HFO": ["PJ/PJ"],
    "LFO": ["PJ/PJ"],
    "crude_oil": ["PJ/PJ"],
    "solar": ["PJ/PJ"],
    "uranium": ["PJ/PJ"],
    "wind": ["PJ/PJ"],
    "electricity": ["PJ/PJ"],
    "heat": ["PJ/PJ"],
}

commodities = units.copy()
for item in ["ProcessName", "RegionName", "Time", "Level", "crude_oil", "biomass", "coal", "LFO", "HFO", "gas"]:
    commodities.pop(item)

for key, _ in commodities.items():
    projections[key] = 0

projections

units = {"RegionName": ["Unit"], "Attribute": ["-"], "Time": ["Year"]}
for commodity in fuels + list(commodities.keys()):
    units[commodity] = ["$/GJ"]
    
units

units_row = pd.DataFrame.from_dict(units, orient="columns")
projections_out = units_row.append(projections)
projections_out
