==> notebooks/converting/8.0-ajmk-checking-outputs.ipynb <==
import pandas as pd

global_commodities = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/processed/starter-kits/Kenya/input/GlobalCommodities.csv")
global_commodities

power_techno = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/processed/starter-kits/Kenya/technodata/power/Technodata.csv")
power_techno

comm_out = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/processed/starter-kits/Kenya/technodata/power/CommOut.csv")

comm_out_unique = list(comm_out.columns)
comm_out_unique

power_unique = list(pd.unique(power_techno.Fuel))
print(power_unique)
commodities_unique = list(pd.unique(global_commodities.CommodityName))
commodities_unique


list(set(commodities_unique) - set(power_unique))

list(set(comm_out_unique) - set(power_unique))
