==> notebooks/converting/10-ajmk-technodata-timeslices.ipynb <==
import pandas as pd

example_ttslices = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/external/muse_data/default_timeslice/technodata/power/TechnodataTimeslices.csv")
example_ttslices

technodata = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/processed/starter-kits/Kenya/technodata/power/Technodata.csv")
technodata

technodata_data = technodata[1:]
technodata_data[['ProcessName', "RegionName", "Time", "Level"]]

pd.unique(technodata_data.ProcessName)

capacity_factors = pd.read_csv("/Users/alexanderkell/Documents/SGI/Projects/11-starter-kits/data/interim/timeslices/Kenya-CFs.csv")
capacity_factors = capacity_factors.melt(id_vars="ProcessName")
capacity_factors.variable = capacity_factors.variable.str.lower()
capacity_factors[['season','day']] = capacity_factors['variable'].str.split(' ',expand=True)
capacity_factors = capacity_factors.drop(columns="variable")
capacity_factors.ProcessName = capacity_factors.ProcessName.str.replace("\bWind\b", "Onshore Wind")
capacity_factors.ProcessName = capacity_factors.ProcessName.str.replace("PV", "Solar PV (Utility)")
capacity_factors = capacity_factors.rename(columns={"value": "UtilizationFactor"})
capacity_factors


technodata_timeslices = pd.merge(technodata_data[['ProcessName', "RegionName", "Time"]], capacity_factors, on="ProcessName")
technodata_timeslices['MinimumServiceFactor'] = 0
technodata_timeslices

process_timeslice = pd.merge(technodata_data[['ProcessName', "Time"]], capacity_factors[['season', 'day']], how="cross").drop_duplicates()
technodata_timeslices = pd.merge(process_timeslice, capacity_factors, on=["ProcessName", "season", "day"], how="outer").set_index(["ProcessName", "Time"]).combine_first(technodata[["ProcessName", "Time","UtilizationFactor"]].set_index(["ProcessName", "Time"])).reset_index()#.to_csv("~/Desktop/result.csv")
technodata_timeslices['MinimumServiceFactor'] = 0
technodata_timeslices
